requirements.txt
# ==========================================
# JARVIS - Assistente Virtual Inteligente
# Versão: 3.0.0
# ==========================================

# === Core ===
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3

# === IA / LLM ===
openai>=1.0.0
anthropic>=0.18.0

# === Cache Semântico (FAISS) ===
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.24.0

# === Pesquisa Web ===
duckduckgo-search>=4.0.0
wikipedia>=1.4.0
tavily-python>=0.3.0  # opcional - pesquisa para IA
requests>=2.31.0

# === Voz (Speech) ===
# Para STT (Speech-to-Text)
# openai-whisper>=20231117  # Whisper local (requer ~1GB)
# faster-whisper>=0.10.0    # Alternativa mais leve
SpeechRecognition>=3.10.0   # Alternativa simples

# Para TTS (Text-to-Speech)
pyttsx3>=2.90              # TTS offline (gratuito)
# elevenlabs>=0.2.0        # TTS premium (opcional)

# Para captura de áudio
PyAudio>=0.2.13            # Windows: pip install pipwin && pipwin install pyaudio
sounddevice>=0.4.6         # Alternativa ao PyAudio

# === Controle do Sistema ===
psutil>=5.9.0              # Info do sistema (CPU, RAM, etc)
pyautogui>=0.9.54          # Automação (opcional)

# === Database ===
mysql-connector-python>=8.0.0
SQLAlchemy>=2.0.0

# === WhatsApp (Selenium fallback) ===
selenium>=4.15.0
webdriver-manager>=4.0.0
undetected-chromedriver>=3.5.0

# === Utilitários ===
colorama>=0.4.6            # Cores no terminal Windows
rich>=13.0.0               # Interface rica (opcional)

# === Validação e Schemas ===
pydantic>=2.0.0            # Validação de dados e schemas

# === Circuit Breaker e Rate Limiting ===
# Nota: Implementação customizada incluída, mas pode usar bibliotecas externas
# pybreaker>=1.0.0         # Circuit breaker (opcional)
# limits>=3.0.0            # Rate limiting (opcional)

# === Dashboard Web ===
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0

# === Sentimento (opcional - fallback sem biblioteca) ===
# vaderSentiment>=3.3.2
# textblob>=0.17.0
